+++ tests/jballerina-unit-test/src/test/resources/test-src/record/record_assignability_negative.bal
@@ -14,8 +14,22 @@
 // specific language governing permissions and limitations
 // under the License.
 
-function testRecordAssignabilityNegative() {
-    record {int|string|boolean a; int|boolean b;} r = {a: 1, b: true};
-    record {int|string|boolean a; int|string b;} r1 = r;
-    record {int|string|boolean a; int|string b;} _ = r;
+function testOpenRecordAssignabilityNegative() {
+    record {int|string|boolean a; int|boolean b;} r1 = {a: 1, b: true};
+    record {int|string|boolean a; int|string b;} r2 = r1;
+    record {int|string|boolean a; int|string b;} _ = r1;
+}
+
+function testClosedRecordAssignabilityNegative() {
+    record {|int|string|boolean a; int|boolean b;|} r1 = {a: 1, b: true};
+    record {|int|string|boolean a; int|string b;|} r2 = r1;
+    record {|int|string|boolean a; int|string b;|} _ = r1;
+
+    record {|int|string|boolean a; int|boolean...;|} r3 = {a: 1, "c": true};
+    record {|int|string|boolean a; int|string...;|} r4 = r3;
+    record {|int|string|boolean a; int|string...;|} _ = r3;
+
+    record {|int|string|boolean a; int|boolean b; int|boolean...;|} r5 = {a: 1, b: true, "c": true};
+    record {|int|string|boolean a; int|string b; int|string...;|} r6 = r5;
+    record {|int|string|boolean a; int|string b; int|string...;|} _ = r5;
 }
